requirements.txt
google-generativeai
pcapy
dpkt
python-dotenv